[feat] set init probe pos asylum
We are not *really* reading the probe position with asylum. It would be
more accurate to say that we are reading the value we will tell it to
move to. Because of this, we might as well set it to a reasonable
'default', i.e. what it would be at if you were manually prepping a
spec.

NOTE: We may change this, especially if we figure out a proper way to
query the position.