[feat] add logic for controller without SS_MOVING
Some controllers do not have a 'moving' event to detect. For these, we
need to 'fake' the event in order for our state machine to match. To
facilitate this, we are:
a) adding a _handle_sending_fake_move() method to MicroscopeTranslator.
b) updating ConfigTranslator to automatically use this if its member
variable detects_moving is False.
c) adding a float_tolerance member variable to MicroscopeTranslator.
This is used by checks_equal(), used by _handle_sending_fake_move().

Since the Asylum controller fits this criteria, we modify its
constructor to (a) set its float tolerance, and (b) tell its parent
classes that it does not detect moving.