[feat] add drift estimation logic
Adding logic for estimating drift (and generally, transforms between
images). This logic is in drift.py, with associated unit tests in
test_drift.py.

Note that the logic here is based heavily on features in skimage, so we
need to add this as a required package. We also add two sample images
for testing this, and update the README to indicate this requirement.

We also update the pyproject/poetry.lock to download gxsmread properly.
For some reason, it does not detect the latest version (0.2.3 :/). I do
not understand why.