[feat] add (unused/broken) check for protobuf
This method is supposed to check whether a given field tracks presence.
I'm including it in the history just in case. However, it (a) does not
work fully, and (b) there is a better way around the problem I was
having.

Essentially, I was trying to clone a protobuf Message recursively, while
changing only a few of its fields. In principle, I only wanted to change
fields that had been set (i.e. their presence indicated True). Going
this route (with this check) does not work consistently.

The better way to do it is to make a new Message of the same type, with
only the field(s) I want to change in it. Then, I can call
old_proto.MergeFrom(new_proto) to merge only those fields.