[fix] change default angle for scan_params_2d
In our helper create_scan_params_2d() method, we were defaulting the
angle to 0 instead of None. This meant that the angle was *set* (i.e.
it's presence was True). This is not great: we want to only set fields
if we explicitly want to do so - i.e. our defaults should not set any
protobuf file.

This CL fixes the situation by (a) defaulting the angle param to None,
and (b) updating this for a couple of associated methods tied to samples.