[feat] update drift/scheduler diff spatial res
Significant changes to the scheduler in drift to account for different
spatial resolutions.
- We add extra logic to get_latest_intersection() to provide a minimum
spatial resolution ratio (difference in spatial resolutions as a ratio).
This is also added to the CSCorrectedScheduler class, with a member
variable for that value.
- In compute_correction_info(), adding logic to extract and scale
patches so they are the same spatial resolution and the common
intersection of both images. This exists in extract_patch() and
extract_and_scale_patches().
- Add tests for this new logic.