[fix] major fix in drift estimation!
The estimation was flipped between x and y axes, because the keypoint
extractor extracts the points as (y, x). However, the Transform that we
estimate functions on vectors as (x, y). Thus, our estimations were
correcting for the wrong axes! We fix this by flipping the keypoints
to (x, y) before estimating our transform.