[feat] major changes to drift
A number of noteworthy things:
- In estimate_transform(), we now output a fitting score in addition to
the mapping. This should allow better determination of whether or not
the fit worked well.
    + The score is also printed in the display method.
- In estimate_transform, we are also adding a scale_factor that the
scans are downscaled to before estimation. This can help lower the
computational cost, as well as minimize noise in the found
keypoints (though a blur would probably work as well for that).
- We added a TRANSLATION TransformType. This essentially just does a
mean on the deviation between matched keypoints. Simpler, and
potentially all we really need. We default to this one now.
- For RANSAC, we have added helper methods to define the kwargs. Of most
use is residual_threshold. This used to be defined in pixel coordinates,
but now is defined as a percentage of the image resolution. Thus, the
default residual threshold is 5% of the image resolution. To simplify
this, we added a method in DriftModel.
      + We also added a helper to define the minimum number of points
      needed to fit for RANSAC (3 for all except
      TransformType.TRANSLATION, which is 2).
- DriftModel now also holds a scan_resolution parameter. This allows us
to update the RANSAC parameters if the resolution changes.
- We also added a fill_nan argument, as the fitting fails if
DescriptorType.BRIEF is used with nan values.
- For all of the above, we update test_drift. We added simulated drifts
in addition to the real data ones.