[fix] error when deleting scans in viz
In our point_subscan example, we were clearing the sub-scan data every N
scans. At least, that was the goal! However, our logic was flushing all
Scan2d messages from the cache. This meant that it was deleting the
full-scans as well. The net result is that the full scan would not
always be shown on updates.

This changes the logic so that we flush all things in the cache that
match a subkey str.