[refactor/fix] fix issues with CorrectionInfo
There were some issues with our DriftSnapshot to CorrectionInfo logic.
First of all, we are clarifying the method names:
- correction_from_drift() becomes estimate_correction_from_snapshot()
- estimate_correction() becomes estimate_correction_from_history()
[NOTE: we pushed part of this change in the prior CL...I'm sorry].

For estimate_correction_from_snapshot(), we update the logic to consider
temporal overlap between the current CorrectionInfo and the provided
DriftSnapshot. It is possible that the snapshot used has matched the
latest scan to a prior scan that is *before* the correction. Thus, we
need to remove the drift associated with that proportion of time.