[fix] get tracking working CSScheduler
Main issues:
0. Having a history of DriftSnapshots doesn't make sense! Each snapshot gives
you the delta in correction from *the current* drift rate and vec. So it is
non-trivial to average them. Instead, we remove this and only use the current
DriftSnapshot to consider things.
1. estimate_correction_no_snapshot() was averaging drift rates from prior
snapshots, which doesn't quite work (or at least not trivially). For now, we
simply consider the current drift rate.
2. estimate_correction_from_snapshot() was not computed properly. Particularly,
it did not account for the fact that a prior drift rate may have been used to
consider the newest scan. We need to consider that contribution when we
calculate the actual vector and actual drift rate!
3. We were *doubling* our drift rate calculation between tot_corr_rate updates.
Particularly, we were using the drift rate to update our vector both in the
control route and in the pubsub route. Really, we only need to be updating in
the control route.