[feat] fir drift estimation edge case
When going from large region/low res scans to small region/high res
ones, we were failing at estiamting drift for (high res, low res) pairs.
Upon inspection, it was because we were always scaling our estimation
inputs to the resolution of the second value in the pair. We were doing
this because the ultimate transform applies to the second pair.

The issue was that we were *downscaling* in these cases. And, for
reasons (not entirely sure), the estimation was failing on these low
resolution images.

The solution is to always scale to the higher resolution of the two. We
now output the scaling done on the second of the two, in order to
counteract later when we estimate the translation.