[feat] update cs scheduler pubsub path correction
Previously, we had decided to *stop* correcting the data being published
by the microscope to go from PCS to SCS. We handwaved it as not being
necessary, as the correction 'had already been done in the control
route'. This was wrong, and I should feel dumb for using that
justification.

What it appears was *really* happening is that the data was being
corrected (unwittingly) due to particularities of how protobuf Messages
are passed around in Python. These data structures are passed around
such that we have references to the original. Thus, in the
CSCorrectedScheduler we were doing something like:
    new_scan = cs_correct_proto(new_scan, ...)
where new_scan was the provided protobuf message. We were in
fact *correcting* the data without realizing it.

Because I find this unclear, I am updating the logic so that we
explicitly update it in CSCorrectedCache and not in
CSCorrectedScheduler. We accomplish the latter by doing a forced
deepcopy of the input proto before we do any calculations. We return to
the originally proposed logic of how/when we update the cache for the
coordinate system mapping.