[feat] add rescanning logic to scheduler
We add logic to rescan in drift/scheduler.CSCorrectedScheduler, which is
to be used in association with drift/rescanner.DriftRescanner. The
scheduler measures the intersection area / rect area between the desired
rect and the true rect (i.e. after drift correction). If below a
threshold, it sends ScanParams out via its publisher. If the
DriftRescanner is subscribed, it will then take control (flagging an
EP_THERMAL_DRIFT problem), rerun the scan, and then release control.