[fix] bug fixes for CSCorrectedScheduler
1. We were having issues with the proto being modified and needing to
make deep copies. The source of this was cs_correct_proto() itself.
Thus, we do a deep copy at the start of it, and remove the other
deepcopies throughout the code.
2. When we update CSCorrectedRouter, we were updating the vector
but *not* the drift rate! Thus, the drift rate was going in the wrong
direction. Updating both to be negative now.