[feat] CSCorrectedScheduler filters scan channel
I realized that CSCorrectedScheduler was treating all scans as equal
when comparing them! In fact, we need to be analyzing just one channel
of our scans for drift (which will likely be multi-channel). Thus, we
update the logic to take a channel_id input, which *must* be explicited.

Linked to this, I have updated ImageTranslator to save a channel_id as
well.

Lastly, I updated the point_subscan_drift config to use an explicit
channel id.