[feat/fix] logic changes to ScanHandler
In previous CLs we have done the following:
1. added a not_in_control check for _handle_scope_state_receipt().
2. Set next params in certain scenarious in handle_issues.

For (1): this actually adds inconsistencies in how the logic was
functioning. In fact, the original ScanHandler logic does not really
concern itself with being in control. This is implicitly handled by
client.py::send_req_handle_ctrl.
For (2): this was adding multiple places where we set self._next_params,
with different 'refresh' rates (handle_issues() vs.
on_message_received()).

What we did:
- We removed both of these. For (2), we moved setting _next_params()
into _perform_scanning_logic(), which is called by both refresh paths.
- We modified handle_rerun to not take in a boolean which wasn't being
used anyway. We also modified the method to (a) update the desired scope
state to moving, and (b) updated the logic to set _next_params to None
if flush_params_on_failure is true.
- We added a new param, flush_params_on_failure. It defaults to False.
The idea is that you can set it to True if the parameters you are
setting are time-specific. Since we cannot guarantee when we will get
control again, this parameter allows us to force a flushing of the last params.
- We enhanced the unit tests for this to try and handle all paths.