[feat] significant changes to pub-sub cache logic
The prior cache logic required exact envelopes to be provided for the
proto-to-history map. If a proto outside of this was provided, the logic
either did not know what to do or would just use a history of 1.

The bulk of the changes here are to allow users to define basic histories,
and for the cache logic to find a 'closest match' when an unprovided envelope
is received.

The simple justifying example: I may choose to store 'Scan2d_20.0_' envelopes
with history 5 (i.e. Scan2ds with a physical scan size of 20 units - we assume
these are consistent). However, the Scans I receive will be channel specific,
e.g. 'Scan2d_20.0_Topo' and 'Scan2d_20.0_Amplitude'. Without the changes here,
these channel-specific envelopes will be given a history of 1 instead of the 5
I desired for all channels.

The bulk of the changes are tied to adding a method get_closest_match() to
compare envelopes. However, we also modified extract_proto() and update_cache().
For the former, we check for the exact proto and if it doesn't exist we look for
the closest match. Since this is used to create an empty proto, we need to find
*any* envelope that matches our type. For the latter, if we do not find the
exact envelope in envelope_to_history_map, we look for the closest match as
well. This is used to to create our initial cache dequeue for that envelope.
Note that if the matching logic fails (which should not happen), we raise an
error.

Making the above changes allowed us to remove the update_cache() overload in
PBCScanLogic. For that class, we also modified its constructor (to function
as originally intended - i.e. adding base scan/spec if not there), and changed
get_envelope_for_proto() to write Scan2ds as Scan2d_PHYSICAL-SIZE_CHANNEL,
with those two variables left blank if not specified by the proto.