[feat] update config.tomls given env changes
Particularly: we explicit our desired channel_id for visualization and
CSCorrectScheduler-ing. For caching, we are using channel-free info,
as we want all channels to be stored with the provided
histories. (Thinking on this, I'm not sure we explicitly need this. In
principle, you could choose to only have the Topo channel have a large
history, if you are sure you are only using that channel for your analysis.)