[feat/fix] misc fixes CSCorrectedScheduler logic
test_components_logic has been failing with a CSCorrectedScheduler. There were
a number of issues found:
- We were not handling the case of self.display_fit being False properly.
In particular, we were using/creating matplotlib objects even if display_fit was
False.
- We were not handing CorrectionInfo being None anywhere. We worked arond this by
setting our default/starting CorrectionInfo to all 0s with a default unit. But
this introduced other weirdnesses. We fix this by...accounting for None cases.
- We were being unclear about CorrectionInfo units. By defaulting to some random
set, we were forcing the units to always be this, even if they were not! Now, we
start with a None (instead of a CorrectionInfo), and the first CorrectionInfo is
created using the same units as the DriftSnapshots (which is to say, the spatial
units of the images compared). All subsequent updates to CorrectionInfo use the
prior CorrectionInfo, so these will always be in consistent units during an
experiment.
    + Linked to this: we *do* do a conversion/check in cs_correct_proto. This is
    because we cannot guarantee that the spatial units of all Message types are
    the same. Thus, we convert from our CorrectionInfo units (which matches the
    spatial units) to those of other Messages we change. In practice, I think
    we are only really correcting ScanParameters2d/Scan2d/Spec1d/ProbePosition,
    which should all be consistent. But it doesn't hurt to be anal about it.