[refactor] simplify minor logic drift/scheduler
One of the 'check for None' updates from the last CL was less clear than
it could have been. Particularly, for deciding when we print out the
PCS-to-SCS correction vector. The 'simpler' logic is simply to check if
(a) we are dealing with our first CorrectionInfo (the main is None, and
the latest is not None), or (b) we have both a current and prior and
they are different.