[fix] get monitor tests running again
It turns out that these tests were failing primarily because our
default SPAWN_DELAY_S in common.py was too short! We had it at 0.25s,
but it needs to be at 0.5s.

This is pretty high, right? My justifications for why we need to do this
are:
- Our IO-based _STARTUP_SLEEP_S, which is used on all IO components, is
already 0.25s! Looking at this, I'm kind of shocked we were passing to
begin with. If all IO subcomponents of a component sleep 0.25s before
returning from the constructor and our monitor only waits 0.25s...you
would think we would run into issues often.
- In order to allow component's to define their own default spawn
delays, we import each class in the monitor's process *before* we spawn
the component-specific process. So this means that we actually import
each class 2x now! So our startup time should be longer, maybe as much
as 2x?

So it seems we may just have to live with these slower startup
times (and, linked to this, longer time for all UTs to run). Oh well.