[test] remove scifireaders dep test_drift
Because we were using scifireaders to load images in test_drift, it was
dependent on it. We don't want this, because depending on the microscope
used by the user, they may not need scifireaders!

Thus, we loaded the files and resaved them as NetCDF files, which are
easily legible by xarray.