[fix] fix mutable default value issue numpy array
For dataclasses, you cannot provide defaults to mutable fields the
default way. It turns out that you have to wrap your default in:
        field(default_factory=lambda: YOUR_DEFAULT)
Do to implementation particularities (look up the Python documentation
on dataclasses, particularly the 'Mutable Default Values' section).

I'm not sure why this wasn't an issue on Linux. The problem only appears
to prop up on Windows?