[fix/feat] add tests to correction / proto geo
Adding/moving tests to correction.py and proto_geo.py. In doing so, I
found a couple of bugs in correction! Particularly, it seems that how we
were doing some of the time overlap calculations to get proper drift
rates was using the wrong time (dt2 instead of dt1), and we were not
considering the div by 0 errors properly.

We're also disabling the test_scheduler tests until they are fixed.