[feat/fix] do not send data units for scan_params
We were using data units in on_set_scan_params() to try and convert. But
at this granularity (i.e. for these calls), the translator has no clue
what units the data are in. So a conversion can't really happen, and
trying to do one (the generic behavior) will cause a crash.

Thus, we are updating the pydocs to explicit this state, and not feeding
data units.