[feat] update ScopeState logic and linked pydoc
I realized that how we handle SS_MOVING is inconsistent! Particularly:
most translators that can detect SS_MOVING events will only detect them
when the probe *actually* moves. If we resend the same scan parameters,
the probe is already where it was and does not move. Thus, we do not get
an SS_MOVING event.

The issue is that this is quite inconsistent. It means that the state
machine we are dealing with has a history: sometimes it will send
SS_MOVING and sometimes it will not. This is terrible for interacting
with.

To 'fix' this, we enforce that all translators should always send a
state SS_MOVING when they accept a probe pos or scan params change,
regardless of whether they actually moved the probe or not. We also
explain the suggested ways to implement this logic, based on whether or
not your microscope tracks SS_MOVING, and in what way it does.

Lastly, we update gxsm to support this logic: now it (a) uses
_handle_sending_fake_move() whenever a scan params or probe pos is
received; and (b) just repeats the prior state when an actual SS_MOVING
is detected. Hopefully, this works. We may end up having to make it skip
a couple of polls  after a request if the logic is not quite right.