[feat] update proto_geo and scheduler scan match
Previously, our matching procedure only grabbed the latest scan in the
cache that matched our spatial intersection and spatial resolution
criteria. This should work well if (a) the surfaces are constant,
and (b) the time between scans is low. However, this is not always the
case. Thus, we add an option to choose between grabbing the 'oldest'
match or the 'latest' match when doing our drift estimation.

Note also that we changed the logic so that now we iterate through all
scans in our cache (that match our criteria), grabbing the first that we
are able to estimate from. Before this, we only grabbed the latest and
if the fit failed we immediately jumped to using the estimated drift
rate. This approach should be more robust, at a potential computational
overhead (more drift estimations per scan).