[feat] udpate io to log with component id
Modifying all the IO subcmponents to take a uuid str, which we use to
feed the component using them. This allows easier reading of logs.

Note that we are adding _uuid and set_uuid() for some IO subcomponents
that (under current configurations) should only really be used by one
component. For example, the router and cache.