[feat] allow per-component default beat periods
This is similar to what we did previously with spawn_delay_s.
Essentially, we are allowing a per-component definition of beat delay,
to account for components that are slow to respond. Similar to
spawn_delay_s, it can be done by overriding DEFAULT_BEAT_DELAY_S; the
rest is handled by the logic in this CL.

We do the same as for spawn_delay_s:
- We create a class method to query DEFAULT_BEAT_DELAY_S.
- We add a method get_component_beat_period_s() in the monitor, to allow
us to query for the proper beat period and feed it to the
HeartbeatListener (otherwise, we would crash on startup because the
listener would expect beats much more quickly than the reality).

We cleaned a couple of things linked to this:
- We removed the instance variable spawn_delay_s, as this is not really
used.
- We updated AfspmComponent's __init__() to have spawn_delay_s and
beat_period_s be None by default. It then classmethod queries for the
proper default delay if either is None (i.e. if it was not explicited by
the user).
- We had to fix MicroscopeTranslator's __init__, as it was overriding
the parent class behaviour (which is what we wanted).