[fix] major bug in drift._estimate_transform
Something in our DriftModel *cannot* be reused, or our model stops
fitting! In general, I prefer to deep copy data that is not duplicated
by default, to avoid these kinds of things.

This CL forces a deepcopy of our DriftModel, to remove the bug. In the
future, I may try to investigate what particular component of DriftModel
is causing this issue.