[fix] crash fiddling ui drift scheduler
If you go crazy with the matplotlib figure for CSCorrectedScheduler, you
can get into a weird crash. It appears this is due to the matplotlib
backend getting stuck, which causes our heartbeats to not be sent (and
us to thus assume the scheduler has frozen). To avoid this, I have
simply increased the default beat period.

NOTE: I have not confirmed this, but rather assume it from testing. I
may find I was wrong and need to revert this change :(.