[refactor] remove CM_PROBLEM from ControlMode!
This is a somewhat big change:we are removing CM_PROBLEM as a
ControlMode. There was a mixing of *problems* and *control modes* in
having CM_PROBLEM as a ControlMode. I tried to fix this previously, but
didn't do a full separation.

In this CL, we simplify the logic:
- Your ControlMode is either AUTOMATED or MANUAL. If in MANUAL mode, no
automation can occur! If in AUTOMATED, a component can take control if
it indicates that it solves a problem that is currently in the problems
set (EP_NONE if it is a general component).
- If a problem is logged, the in-control component loses control. If
they can resolve this new problem, they can re-request control.
- If we switch to MANUAL, the in-control component loses control. No
component can gain control until we switch back to AUTOMATED.

We updated control_ui accordingly, feedback.py, and various tests.
We also added a test for the specific case that was broken (which made
us do this refactor): if in MANUAL mode with a flagged problem, an
automated component could take over! No bueno.

Note that there was also a bug in router.py:147, where I had !A and B or
C rather than !A and (B or C). (I suppose the not is causing weirdness?).