[feat] add scripts for drift analysis
Adding two useful scripts for drift analysis:
- offline.py: this allow us to rerun a cruder form of drift estimation
on data where we did not run CSCorrectedScheduler. The main limitation
is that it only uses a cache size of 1, so the estimates may be less
good.
- visualize.py: this allows us to visualize the drift offsets and rates
for an experimental run (via the drift_correction.csv file).

Note also that we have made a 'scripts' folder, to hold this kind of
info. We have chosen to organize not *exactly* like afspm. In this case,
it doesn't necessarily make sense to put the drift folder in a
components subfolder (as there aren't components in scripts?). I reserve
the right to change my mind on this, though!