[fix] scaling drift_rates size properly
Earlier, you were using units=xy. Apparently this scales as a function
of the size of the vector, so you were getting inconsistent results
between experiments! You are switching to 'width', which scales relative
ot the figure size (I think). Note that the 'width' magic number needed
to change to account for this. Now it seems to be consistent between
datasets at least.