[fix] update scifireaders data to match asylum cs
It appears that the scifireaders IgorIBWReader improperly reads the
scans, swapping the axes when reading. To work around it for now, we are
re-swapping them back. The more proper fix is to correct this directly
in the package. I will do this.

Once this is fixed on the scifireaders end, we can undo this workaround.