[refactor] make correction logic more sensible
Previously, we were negating the drift snapshot, because we want to
correct for it. However, this is confusing: we should simply be negating
the drift snapshot when we *calculate* the correction. Part of this CL
involves fixing that.
The other part relates to the usage of 'drift_rate' and 'drift_offset',
when I meant 'correction_rate' and 'correction_offset'. I am trying to
remove references to drift rate unless it *actually* refers to the drift
rate.
I also updated test_correction to pass with the previous CL's changes,
where we removed the 'time overlap consideration' part of our correction
estimation from a drift snapshot.