[feat] change keep_scanning logic
We used to try and grab the latest scan params and redo that. However,
the latest scan params will be updated by the estimated drift, and we
know that there will be scans where there is drift (and thus the latest
scan params are not our original desired ones). Thus, we switch now to
expliciting a scan region, and just constantly requesting that.