[feat] updates to metadata to better store state
From running experiments, I realized that the person in control was
rarely being well stored by metadata.py. This CL adds a minor state
machine to update the control state (only doing so when we are in
SS_SCANNING or SS_SPEC), and updates the csv to be a little easier to
read (timestamps become datetimes, both the ControlMode and
ExperimentProblem enums become strings).

Note that we store the list of problems as PROBLEMA-PROBLEMB, i.e.
separated by a '-' to avoid conflicts with ',' for CSVs.