[fix] (maybe very big) in asylum translator
It looks like the scans *also* do not store the origin (this was found
to be the case with probe_pos). So we need to store the latest scan
params from when a scan was requested as well, and update the scans
after our reader loads them.

If this is the source of issues, it is a *huge* bug I did not notice
until now. It really begs the question: how was *anything* working
earlier if this is the case?