[feat] remove fake_correction option offline.py
It cannot really be done properly (or would be pretty involved), as we
need to update the scan params coordinates to be the PCS coordinates
that *would have been* sent. So I am removing it. If you want to
'emulate' the experiment, just set the cache size to larger than the
total scans.