[feat/refactor] updates to test_translator
Mainly, we are introducing two new tests to test_translator:
- test_scan_coords ensures that the final scan we read has the
appropriate physical coordinates we set it with. This is useful because
some microscopes save scans *without* their coordinates in the piezo
range. The translator is thus expected to store this information and
update the Scan2d's parameters with the proper info.
- test_spec_coords does the equivalent for Spec1ds (ensuring the probe
position is properly stored).
- We do some minor refactoring (making common methods) to avoid a
too-huge test (although you could argue they are still too big :( ).

NOTE: it is very likely I have introduced bugs! There will likely be a
'fix' CL shortly after this, after I have tested it on one of the microscopes.