[feat] working but very hacked keypoints script
This script is almost illegible, I hate it. But it works, so I am
keeping it for now. It allows us to save images of two scans with
keypoints detected and visualized, and the inliers and outliers shown.

If you need a colorbar, remove the 'add_colorbar=False' and
'plt.gca().set_aspect('equal')' calls.