[feat] add logic to draw pixel units in figures
This is done (for now) only on the exclusive X- and Y- graphs below the
main X/Y one. If we end up wanting it in the other one too, you will
need to figure out what is going on when you do ax.twinx().twiny(). It
does not seem to allow adding labels for both sides...