[feat] switch to AGPL 3.0 license
Here, we are switching to the Affero GPL license. This (in principle)
forces users of our framework that are using it over a netwrok to share
the code in an equivalent manner to GPL.

I could have modified history to have this from the beginning, but this
seems insincere. I hope features added after this will convince users to
use versions of afspm with the AGPL license.